Finished Client
Can now grab data from both API
