+++ No-Docker/node-module/server.js
@@ -1,9 +1,12 @@
 const { PostgresClient } = require("./postgresClient");
+const cors = require("cors");
 const express = require("express");
 const app = express();
 
+app.use(cors());
+
 app.use(function (req, res, next) {
-  var whitelist = ["localhost:5000", "localhost:3000"];
+  var whitelist = ["localhost:5000", "http://localhost:3000"];
   var host = req.get("host");
 
   whitelist.forEach(function (val, key) {
@@ -15,7 +18,7 @@ app.use(function (req, res, next) {
   next();
 });
 
-app.get("/", async (req, res) => {
+app.get("/node-api", async (req, res) => {
   const client = new PostgresClient();
   const response = await client.SelectAllQuery();
   res.json(response);
