+++ No-Docker/react-module/web-client/src/App.js
@@ -2,16 +2,10 @@ import DataFetcher from "./Components/DataFetcher";
 import { apiRoutes } from "./Networking/apiRoutes";
 import { baseUrl } from "./Networking/baseUrl";
 import { Ports } from "./Networking/ports";
+import { styles } from "./Styles/styles";
 function App() {
   return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "row",
-        width: "100%",
-        justifyContent: "space-evenly",
-      }}
-    >
+    <div style={styles.dataContainer}>
       <div>
         <h3>DATA FROM PYTHON API</h3>
         <DataFetcher
@@ -21,8 +15,12 @@ function App() {
         />
       </div>
       <div>
-        <h3>DATA FROM NODE JS</h3>
-        HELLO
+        <h3>DATA FROM NODE API</h3>
+        <DataFetcher
+          url={baseUrl.url}
+          port={Ports.nodePort}
+          route={apiRoutes.node_api}
+        />
       </div>
     </div>
   );
