+++ No-Docker/react-module/web-client/src/Networking/apiRoutes.js
@@ -1,3 +1,4 @@
 export const apiRoutes = {
   python_api: "/python-api",
+  node_api: "/node-api"
 };
