+++ No-Docker/react-module/web-client/src/Networking/ports.js
@@ -1,3 +1,4 @@
 export const Ports = {
   pythonPort: 5000,
+  nodePort: 8765,
 };
