--- No-Docker/react-module/web-client/src/Styles/DataContainer.js
@@ -1,9 +1,0 @@
-
-export const styles = {
-    dataContainer : {
-        display: "flex",
-        flexDirection: "row",
-        justifyContent:"space-evenly",
-        width: "100%"
-    }
-}
