+++ No-Docker/react-module/web-client/src/Styles/styles.js
@@ -1,0 +1,8 @@
+export const styles = {
+  dataContainer: {
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-evenly",
+    width: "100%",
+  },
+};
